fix number issue

+++ lib/forestvpn_cloud_api/models/create_or_update_data_usage_limit_request.rb
@@ -38,7 +38,7 @@ module ForestVPNCloudAPI
     def self.openapi_types
       {
         :'duration' => :'String',
-        :'value' => :'Float'
+        :'value' => :'Integer'
       }
     end
